move ballsdex entrypoint to use django
