+++ admin_panel/preview/utils.py
@@ -1,8 +1,3 @@
-import os
-
-from tortoise import Tortoise
-
-from ballsdex.__main__ import init_tortoise
 from ballsdex.core.models import (
     Ball,
     Economy,
@@ -23,8 +18,6 @@ async def refresh_cache():
     This must be called on every request, since the image generation relies on cache and we
     do *not* want caching in the admin panel to happen (since we're actively editing stuff).
     """
-    if not Tortoise._inited:
-        await init_tortoise(os.environ["BALLSDEXBOT_DB_URL"], skip_migrations=True)
     balls.clear()
     for ball in await Ball.all():
         balls[ball.pk] = ball
